requirements.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
graphviz==0.20.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.4
